bring fetch alerts function to parent component
